+++ lib/DisplayPatternLib/src/ColorPatterns/BackgroundPlusThree.h
@@ -9,12 +9,12 @@
 class BackgroundPlusThree : public ColorPattern
 {
     public:
-        BackgroundPlusThree(ulong background, ulong color1, ulong color2, ulong color3);
+        BackgroundPlusThree(uint32_t background, uint32_t color1, uint32_t color2, uint32_t color3);
 
         virtual void reset();
-        virtual ulong getNextColor();
-        virtual void incrementOnly(uint incrementAmount);
-        virtual uint getNumberOfParameters();
+        virtual uint32_t getNextColor();
+        virtual void incrementOnly(uint16_t incrementAmount);
+        virtual uint16_t getNumberOfParameters();
 
         void setBackgroundDuration(byte duration);
         void setColor1Duration(byte duration);
@@ -24,17 +24,17 @@ class BackgroundPlusThree : public ColorPattern
         static std::vector<String> getParameterNames();
 
     private:
-        ulong m_backgroundColor{0};
-        ulong m_color1{0};
-        ulong m_color2{0};
-        ulong m_color3{0};
+        uint32_t m_backgroundColor{0};
+        uint32_t m_color1{0};
+        uint32_t m_color2{0};
+        uint32_t m_color3{0};
         byte m_backgroundDuration{1};
         byte m_color1Duration{1};
         byte m_color2Duration{1};
         byte m_color3Duration{1};
-        ulong m_iterationCount{0};
+        uint32_t m_iterationCount{0};
 
-        std::vector<ulong> m_colorSequence;
+        std::vector<uint32_t> m_colorSequence;
         byte convertDuration(byte duration);
 };
